renaming the module from 'physical' to 'dynamics'
