+++ src/scores/dynamics/__init__.py
@@ -2,7 +2,7 @@
 Import the functions from the implementations into the public API
 """
 
-from scores.physical.energetics_impl import energy_components_lat_lon, energy_exchanges_lat_lon
+from scores.dynamics.energetics_impl import energy_components_lat_lon, energy_exchanges_lat_lon
 
 __all__ = [
     "energy_components_lat_lon",
