+++ src/scores/dynamics/budgets_utils.py
